Refactor: GET YOUR CODE OFF MY ~~LAWN~~ HEADERS
Refactor FFmpeg and Audio.
