+++ game/audio.hh
@@ -4,17 +4,43 @@
 #include "ffmpeg.hh"
 #include "notes.hh"
 #include "pitch.hh"
-#include "util.hh"
 #include "libda/portaudio.hpp"
 #include <deque>
 #include <map>
 #include <memory>
+#include <mutex>
 #include <string>
+#include <unordered_map>
 #include <vector>
 
 const unsigned AUDIO_MAX_ANALYZERS = 11;
 
 struct Output;
+
+/**
+* Advanced audio sync code.
+* Produces precise monotonic clock synced to audio output callback (which may suffer of major jitter).
+* Uses system clock as timebase but the clock is skewed (made slower or faster) depending on whether
+* it is late or early. The clock is also stopped if audio output pauses.
+**/
+class AudioClock {
+	mutable std::mutex m_mutex;
+	Time m_baseTime; ///< A reference time (corresponds to m_basePos)
+	Seconds m_basePos = 0.0s; ///< A reference position in song
+	double m_skew = 0.0; ///< The skew ratio applied to system time (since baseTime)
+	std::atomic<Seconds> m_max{ 0.0s }; ///< Maximum output value for the clock (end of the current audio block)
+	/// Get the current position (current time via parameter, no locking)
+	Seconds pos_internal(Time now) const;
+public:
+	/**
+	* Called from audio callback to keep the clock synced.
+	* @param audioPos the current position in the song
+	* @param length the duration of the current audio block
+	*/
+	void timeSync(Seconds audioPos, Seconds length);
+	/// Get the current position in seconds
+	Seconds pos() const;
+};
 
 struct Device {
 	// Init
@@ -104,5 +130,37 @@ public:
 	/** Do a pitch shift - used for guitar whammy bar */
 	void streamBend(std::string track, double pitchFactor);
 	/** Get sample rate */
-	double getSR() const { return 48000.0; }
+	static double getSR() { return 48000.0; }
+class Music {
+struct Track {
+	FFmpeg mpeg;
+	float fadeLevel = 1.0f;
+	float pitchFactor = 0.0f;
+	template <typename... Args> Track(Args&&... args): mpeg(args...) {}
+};	
+	public:
+	std::unordered_map<std::string, std::unique_ptr<Track>> tracks; ///< Audio decoders
+	double srate; ///< Sample rate
+	int64_t m_pos = 0; ///< Current sample position
+	bool m_preview;
+	class AudioClock m_clock;
+	Seconds durationOf(int64_t samples) const { return 1.0s * samples / srate / 2.0; }
+	float* sampleStartPtr = nullptr;
+	float* sampleEndPtr = nullptr;
+public:
+	bool suppressCenterChannel = false;
+	double fadeLevel = 0.0;
+	double fadeRate = 0.0;
+	using Buffer = std::vector<float>;
+	Music(Audio::Files const& files, unsigned int sr, bool preview);
+	/// Sums the stream to output sample range, returns true if the stream still has audio left afterwards.
+	bool operator()(float* begin, float* end);
+	void seek(double time) { m_pos = time * srate * 2.0; }
+	/// Get the current position in seconds
+	double pos() const { return m_clock.pos().count(); }
+	double duration() const;
+	/// Prepare (seek) all tracks to current position, return true when done (nonblocking)
+	bool prepare();
+	void trackFade(std::string const& name, double fadeLevel);
+	void trackPitchBend(std::string const& name, double pitchFactor);	
 };
